open gov refactoring - explore overview component
