+++ packages/renderer/src/renderer/screens/OpenGov/Overview/types.ts
@@ -5,3 +5,8 @@ export interface OverviewProps {
   setSection: React.Dispatch<React.SetStateAction<number>>;
   setSectionContent: React.Dispatch<React.SetStateAction<string>>;
 }
+
+export interface OverviewExploreProps {
+  setSection: React.Dispatch<React.SetStateAction<number>>;
+  setSectionContent: React.Dispatch<React.SetStateAction<string>>;
+}
